converter: materialize EVERY untyped constant in an `any` slot at Go's DEFAULT TYPE
Go boxes an untyped constant into an interface at its default type -- untyped int ->
`int` (go2cs `nint`), untyped rune -> `rune` (int32), untyped float -> `float64` -- and
every observation of an interface value dispatches on the boxed CLR type: `x.(int)`
(`x._<nint>()`), `case int:`, golib `AreEqual`'s `leftType != right.GetType()` gate, and
fmt's printArg type-switch. Two renderings diverged.

(1) THE VARIADIC LITERAL FAST PATH. A bare int literal in a `...any` slot was deliberately
left uncast on the theory that a boxed Int32 formats identically to nint under %d/%v, so
the cast was redundant noise on the most common call pattern (fmt/print/log). That is
wrong the moment the box is COMPARED rather than printed: encoding/base32's
`testEqual("Read after EOF, n = %d, expected %d", n, 0)` boxed n as nint and 0 as Int32,
AreEqual compared the dynamic types first and reported them unequal, and the assert fired
with a message that reads as an equality -- "Read after EOF, n = 0, expected 0". It is
also wrong on the %!-verb path, where the full-conversion fmt names the argument via
reflect.TypeOf(arg).String() and a boxed Int32 reports "int32" instead of Go's "int"
(the residual gap 9129d40a1 disclosed). The carve-out is gone; per the user ruling the
churn is accepted, and it turned out to be two lines in two corpus files (Go's own stdlib
almost never passes a bare int literal into a `...any` slot).

(2) THE RUNE AND FLOAT KINDS. 9129d40a1 fixed the named-untyped-INT half; a named untyped
RUNE const (`const Escape = '\xff'`, os.PathSeparator) and a named untyped FLOAT const
both render as the golib UntypedInt/UntypedFloat WRAPPER STRUCT, which matches no Go type
at all, and neither was cast anywhere -- an interface comparison against the same value
reported unequal, a type switch fell through to `default`, and fmt's dispatch dropped to
reflection. `%c` of os.PathSeparator in testing's TempDir/outputDir is a real Phase-4
consumer of this (two corpus lines, both fixed here).

argBoxesAsInt32ButNeedsNint -> untypedConstBoxCast, now returning the C# cast TYPE (or
none) rather than a bool: the constant's KIND comes from info.Types[arg] (the type
go/types already DEFAULTED for the interface slot), while whether the rendering is a
wrapper struct comes from exprRendersUntypedConstWrapper (9129d40a1's predicate, widened to
UntypedFloat) -- info.Types cannot answer that, reporting plain `int` for a literal and a
named untyped const alike. The int32-range test now applies to the LITERAL rendering only,
since a wrapper needs the cast at every magnitude (`const namedWide = 1<<40` in an any slot
was boxing a struct too). boxUntypedIntAsNint -> boxUntypedConstAsDefaultType for the
non-call positions.

SITES COVERED (all measured against the real emission, not assumed): variadic `...any` and
non-variadic `any` call arguments, var-spec, assignment and reassignment, `return`, channel
send, []any/[N]any elements, keyed and positional struct `any` fields, map[K]any VALUES,
map[any] KEYS (composite AND index lookup), the interface ==/!= AreEqual lowering, and an
explicit `any(...)` conversion. The last three were newly wired here:

  - map[any] KEY was a documented deliberate omission -- golib's map uses the default
    Dictionary comparer (nint(6) != Int32(6)) and leaving BOTH the composite key and a
    literal index uncast kept `map[any]int{6:1}[6]` round-tripping. That self-consistency
    only held literal-vs-literal: a lookup by a real `int` VALUE (m[n], necessarily boxed
    nint -- the only form Go can distinguish) MISSED its own stored key. convIndexExpr now
    applies the same cast to an untyped-constant index of an any-keyed map, so store and
    lookup agree while `m[int32(6)]` correctly misses, as Go requires.
  - `any(7)` (explicit conversion to the empty interface) boxed System.Int32, so the paired
    `any(7).(int)` assertion failed. Hops through the default type beside the existing
    `string(...)` hop.

SITES DELIBERATELY LEFT: untyped COMPLEX constants -- a named one renders as golib
GoUntyped (a BigInteger.Parse of the literal text, visitValueSpec's writeUntypedConst path
with its own standing TODO), a separate pre-existing gap a `(complex128)` cast would not
close; a complex LITERAL already boxes correctly. Untyped string/bool need nothing (@string
/ bool are already the Go types). A type-PARAMETER slot constrained by `any` stays excluded
(isEmptyInterfaceTarget) -- instantiation binds the concrete type. A cast onto an expression
that already renders at the default type -- a TYPED constant (`const seqFirst int = iota` ->
`const nint`), or a local untyped const visitValueSpec tightened to a concrete C# type -- is
a harmless no-op left unsuppressed: the predicate cannot know the declaration's chosen C#
spelling from the argument alone, and an extra cast is cosmetic where a missing one is a
runtime divergence.

Gates: check-no-regression 11 behavioral projects drifted, every distinct shape inspected
and classified (bare int literal into a variadic slot: ErrorfFormatting, PrintfWidthFlags,
ShiftPrecedenceUnsigned, VariadicClosureSpread, VariadicFuncFields, VariadicFuncTypeAssert,
VariadicFuncValues; named untyped FLOAT const: ComplexConstContext, UntypedConstDefine;
map[any] int key + lookup: AnyKeyMap; typed-int const no-op cast: IotaEnum) -- all 11
re-baselined via UpdateTestTargets. Full behavioral suite 476/476 across all four phases
(446 Output-compared, 30 skipped). Isolation-reconvert-diff over the full 302-package
corpus, this converter vs the 9129d40a1 converter, CRLF-normalized: exactly 3 files /
4 lines differ (+2 `(nint)(`, +2 `(int32)(` across 1694 files) -- go/token/position.cs,
internal/buildcfg/cfg.cs, testing/testing.cs -- and all three packages build clean with the
fresh emission overlaid. The overlay was restored, not banked (no partial rebank).

Guard: AnyBoxedUntypedConst pins the whole class in one output-compared program -- variadic
and non-variadic slots; named/literal/rune/float/beyond-int32 constants; []any, map[any],
map[K]any, struct-field, chan-send, return and explicit-conversion positions; the map[any]
store<->lookup round-trip plus its int32 miss; and the dynamic types a type switch reports.
Counter-proven: the pre-fix converter diverges on 13 of its lines.

Docs: ConversionStrategies-Reference.md's Empty Interface section rewritten from
"untyped int -> nint" to the general default-type rule (covered sites, the two carve-outs
that were removed and why, the residues that remain); ConversionStrategies.md's summary
follows.

(cherry picked from commit 08700584900989dd2998322d120d8b26986e50af)

+++ src/Tests/Behavioral/AnyKeyMap/AnyKeyMap.cs
@@ -8,10 +8,10 @@ internal static void Main() {
     map<any, @string> tbl = default!;
     tbl = new map<any, @string>();
     tbl["12"] = "1"u8;
-    tbl[12] = "2"u8;
+    tbl[(nint)(12)] = "2"u8;
     tbl[12.0D] = "3"u8;
     fmt.Println(tbl["12"]);
-    fmt.Println(tbl[12]);
+    fmt.Println(tbl[(nint)(12)]);
     fmt.Println(tbl[12.0D]);
 }
